backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
sentencepiece
Pillow
numpy
pandas
requests



